Hide not current l3+ entries (#73)
Hide not current l3+ entries

In the sidebar we have a problem that we keep all tree open just to be
able to permanently show l2 level (services, developer, additional,
etc). This is causing us to also keep all the underlaying trees open as
well (Sphinx just does not let us do it differently). In order to have
possibility to simulate collapse of the not current elements in the tree
add a css hack that will hide subtree of not "current" l2 elements.

Reviewed-by: Vladimir Hasko <vladimirhasko@gmail.com>